+++ examples/Benchmark/Render/ManyFrames.scn
@@ -1,6 +1,4 @@
 <?xml version="1.0"?>
-
-<!-- BeamFEMForceField example -->
 <Node name="root" dt="0.01">
     
     <RequiredPlugin name="Sofa.Component.Engine.Transform"/> <!-- Needed to use components [Vertex2Frame] -->
@@ -22,49 +20,5 @@
             <UniformMass totalMass="0.1" />
         </Node>
     </Node>
-<!-- 
-    <Node name="beam-withTriangulatedCubeCollision">
-        <EulerImplicitSolver rayleighStiffness="0" printLog="false"  rayleighMass="0.1" />
-        <BTDLinearSolver template="BTDMatrix6d" printLog="false" verbose="false" />
-        <MechanicalObject template="Rigid3" name="DOFs" position="0 0 -1 0 0 0 1  1 0 -1 0 0 0 1  2 0 -1 0 0 0 1  3 0 -1 0 0 0 1  4 0 -1 0 0 0 1  5 0 -1 0 0 0 1  6 0  -1 0 0 0 1  7 0 -1 0 0 0 1" />
-        <MeshTopology name="lines" lines="0 1 1 2 2 3 3 4 4 5 5 6 6 7" />
-        <FixedProjectiveConstraint name="FixedProjectiveConstraint" indices="0" />
-        <UniformMass vertexMass="1 1 0.01 0 0 0 0.1 0 0 0 0.1" printLog="false" />
-        <BeamFEMForceField name="FEM" radius="0.1" radiusInner="0" youngModulus="20000000" poissonRatio="0.49"/>
 
-        <Node name="Collision">
-            <CubeTopology nx="15" ny="2" nz="2" min="0 -0.1 -0.1" max="7 0.1 0.1" />
-            <MechanicalObject />
-            <BeamLinearMapping isMechanical="true" />
-            <TriangleCollisionModel />
-            <LineCollisionModel />
-            <PointCollisionModel  />
-        </Node>
-    </Node>
-
-
-    <Node name="beam-withSphereCollision">
-        <EulerImplicitSolver rayleighStiffness="0" printLog="false" rayleighMass="0.1"/>
-        <CGLinearSolver threshold="0.000000001" tolerance="0.0000000001" iterations="25" printLog="false" />
-
-        <MechanicalObject template="Rigid3" name="DOFs" position="0 0 1 0 0 0 1  1 0 1 0 0 0 1  2 0 1 0 0 0 1  3 0 1 0 0 0 1" />
-        <MeshTopology name="lines" lines="0 1 1 2 2 3" />
-        <FixedProjectiveConstraint name="FixedProjectiveConstraint" indices="0" />
-        <UniformMass totalMass="4" />
-        <BeamFEMForceField name="FEM" radius="0.05" radiusInner="0" youngModulus="20000000" poissonRatio="0.49"/>
-
-        <Node name="Collision">
-            <MechanicalObject />
-            <SphereCollisionModel radius="0.4" />
-            <IdentityMapping />
-        </Node>
-    </Node>
-
-
-    <Node name="Floor">
-        <MeshOBJLoader name="loader" filename="mesh/floor3.obj" scale3d="0.5 0.5 0.5"/>
-        <MeshTopology src="@loader" />
-        <MechanicalObject src="@loader"  dy="-1"/>
-        <TriangleCollisionModel name="FloorTriangle" simulated="0" moving="0" contactStiffness="100" />
-    </Node> -->
 </Node>
